major first refactor for cli replacement
