+++ packages/cli/package.json
@@ -37,6 +37,7 @@
   "dependencies": {
     "@as-pect/assembly": "^3.2.0",
     "@as-pect/core": "^3.2.0",
+    "ansi-colors": "4.1.1",
     "chalk": "^4.0.0",
     "glob": "^7.1.6"
   },
