--- packages/cli/src/test.ts
@@ -1,1 +1,0 @@
-require("../lib/index.js").asp(process.argv.slice(2));
