--- packages/cli/src/worklets/ICommand.ts
@@ -1,9 +1,0 @@
-/**
- * @ignore
- *
- * This interface defines a message type for the compiler worklet.
- */
-export interface ICommand {
-  type: string;
-  props: any;
-}
