+++ packages/core/src/reporter/VerboseReporter.ts
@@ -1,5 +1,5 @@
 import { TestContext } from "../test/TestContext";
-import { IWritable } from "../util/IWriteable";
+import { IWritable } from "../util/IWritable";
 import { ReflectedValue } from "../util/ReflectedValue";
 import { TestNodeType } from "../util/TestNodeType";
 import { TestNode } from "../test/TestNode";
