--- packages/core/src/util/IWriteable.ts
@@ -1,7 +1,0 @@
-/**
- * This interface is a utitily used to describe the shape of something that has a `write()` method.
- */
-export interface IWritable {
-  /** This method is used for writing string contents to something that is writable. */
-  write(chunk: string): void;
-}
